fix: sqlite

+++ model/dto/comment.go
@@ -3,7 +3,7 @@ package dto
 import "github.com/go-sonic/sonic/consts"
 
 type Comment struct {
-	ID                int64                `json:"id"`
+	ID                int32                `json:"id"`
 	Author            string               `json:"author"`
 	Email             string               `json:"email"`
 	IpAddress         string               `json:"ipAddress"`
@@ -12,7 +12,7 @@ type Comment struct {
 	Content           string               `json:"content"`
 	Status            consts.CommentStatus `json:"status"`
 	UserAgent         string               `json:"userAgent"`
-	ParentID          int64                `json:"parentId"`
+	ParentID          int32                `json:"parentId"`
 	IsAdmin           bool                 `json:"isAdmin"`
 	AllowNotification bool                 `json:"allowNotification"`
 	CreateTime        int64                `json:"createTime"`
